Move fetchWithAuth to a more appropriate place
